+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/trackselection/DefaultTrackSelectorTest.java
@@ -1521,6 +1521,88 @@ public final class DefaultTrackSelectorTest {
   }
 
   @Test
+  public void selectTracks_withPreferredTextLabel_selectsLabeledTrack() throws Exception {
+    Format unlabeledText = TEXT_FORMAT.buildUpon().setLabel(null).build();
+    Format labeledText = TEXT_FORMAT.buildUpon().setLabel("commentary").build();
+    TrackGroupArray trackGroups = wrapFormats(unlabeledText, labeledText);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_TEXT_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters.buildUpon().setPreferredTextLabels("commentary").build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, labeledText);
+  }
+
+  @Test
+  public void selectTracks_withMultiplePreferredTextLabels_selectsHigherPriorityLabeledTrack()
+      throws Exception {
+    Format lowPriorityLabeledText = TEXT_FORMAT.buildUpon().setLabel("low_priority").build();
+    Format highPriorityLabeledText = TEXT_FORMAT.buildUpon().setLabel("high_priority").build();
+    TrackGroupArray trackGroups = wrapFormats(lowPriorityLabeledText, highPriorityLabeledText);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_TEXT_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredTextLabels("high_priority", "low_priority")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, highPriorityLabeledText);
+  }
+
+  @Test
+  public void selectTracks_roleFlagsPreferredOverTextLabel() throws Exception {
+    Format labeledButWrongRoleFlag =
+        TEXT_FORMAT
+            .buildUpon()
+            .setRoleFlags(C.ROLE_FLAG_SUPPLEMENTARY)
+            .setLabel("commentary")
+            .build();
+    Format unlabeledButRightRoleFlag =
+        TEXT_FORMAT.buildUpon().setRoleFlags(C.ROLE_FLAG_CAPTION).setLabel(null).build();
+    TrackGroupArray trackGroups = wrapFormats(labeledButWrongRoleFlag, unlabeledButRightRoleFlag);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_TEXT_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredTextRoleFlags(C.ROLE_FLAG_CAPTION)
+            .setPreferredTextLabels("commentary")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, unlabeledButRightRoleFlag);
+  }
+
+  @Test
+  public void selectTracks_noSelectionWhenLanguageMismatchDespiteLabelMatch() throws Exception {
+    Format labeledButWrongLanguage =
+        TEXT_FORMAT.buildUpon().setLanguage("deu").setLabel("commentary").build();
+    TrackGroupArray trackGroups = wrapFormats(labeledButWrongLanguage);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_TEXT_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredTextLanguage("eng")
+            .setPreferredTextLabels("commentary")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertNoSelection(result.selections[0]);
+  }
+
+  @Test
   public void selectTracks_selectTextByDefault_selectsTrackWhenNoOtherPreferencesSet()
       throws Exception {
     // A text track with no language, role, or selection flags.
@@ -2613,6 +2695,61 @@ public final class DefaultTrackSelectorTest {
   }
 
   @Test
+  public void selectTracks_withPreferredVideoLabel_selectsLabeledTrack() throws Exception {
+    Format unlabeledVideo = VIDEO_FORMAT.buildUpon().setLabel(null).build();
+    Format labeledVideo = VIDEO_FORMAT.buildUpon().setLabel("commentary").build();
+    TrackGroupArray trackGroups = wrapFormats(unlabeledVideo, labeledVideo);
+    RendererCapabilities[] rendererCapabilities = new RendererCapabilities[] {VIDEO_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters.buildUpon().setPreferredVideoLabels("commentary").build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, labeledVideo);
+  }
+
+  @Test
+  public void selectTracks_withMultiplePreferredVideoLabels_selectsHigherPriorityLabeledTrack()
+      throws Exception {
+    Format lowPriorityLabeledVideo = VIDEO_FORMAT.buildUpon().setLabel("low_priority").build();
+    Format highPriorityLabeledVideo = VIDEO_FORMAT.buildUpon().setLabel("high_priority").build();
+    TrackGroupArray trackGroups = wrapFormats(lowPriorityLabeledVideo, highPriorityLabeledVideo);
+    RendererCapabilities[] rendererCapabilities = new RendererCapabilities[] {VIDEO_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredVideoLabels("high_priority", "low_priority")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, highPriorityLabeledVideo);
+  }
+
+  @Test
+  public void selectTracks_videoRoleFlagsPreferredOverVideoLabel() throws Exception {
+    Format labeledButWrongRoleFlag =
+        VIDEO_FORMAT.buildUpon().setRoleFlags(C.ROLE_FLAG_DUB).setLabel("commentary").build();
+    Format unlabeledButRightRoleFlag =
+        VIDEO_FORMAT.buildUpon().setRoleFlags(C.ROLE_FLAG_COMMENTARY).setLabel(null).build();
+    TrackGroupArray trackGroups = wrapFormats(labeledButWrongRoleFlag, unlabeledButRightRoleFlag);
+    RendererCapabilities[] rendererCapabilities = new RendererCapabilities[] {VIDEO_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredVideoRoleFlags(C.ROLE_FLAG_COMMENTARY)
+            .setPreferredVideoLabels("commentary")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, unlabeledButRightRoleFlag);
+  }
+
+  @Test
   public void selectTracks_withViewportSize_selectsTrackWithinViewport() throws Exception {
     Format formatH264Low =
         new Format.Builder()
@@ -3139,6 +3276,68 @@ public final class DefaultTrackSelectorTest {
     assertFixedSelection(result.selections[0], trackGroups, formatAac);
   }
 
+  @Test
+  public void selectTracks_withPreferredAudioLabel_selectsLabeledTrack() throws Exception {
+    Format unlabeledAudio = AUDIO_FORMAT.buildUpon().setLabel(null).build();
+    Format labeledAudio = AUDIO_FORMAT.buildUpon().setLabel("commentary").build();
+    TrackGroupArray trackGroups = wrapFormats(unlabeledAudio, labeledAudio);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_AUDIO_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters.buildUpon().setPreferredAudioLabels("commentary").build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, labeledAudio);
+  }
+
+  @Test
+  public void selectTracks_withMultiplePreferredAudioLabels_selectsHigherPriorityLabeledTrack()
+      throws Exception {
+    Format lowPriorityLabeledAudio = AUDIO_FORMAT.buildUpon().setLabel("low_priority").build();
+    Format highPriorityLabeledAudio = AUDIO_FORMAT.buildUpon().setLabel("high_priority").build();
+    TrackGroupArray trackGroups = wrapFormats(lowPriorityLabeledAudio, highPriorityLabeledAudio);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_AUDIO_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredAudioLabels("high_priority", "low_priority")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, highPriorityLabeledAudio);
+  }
+
+  @Test
+  public void selectTracks_roleFlagPreferredOverAudioLabel() throws Exception {
+    Format labeledButWrongRoleFlag =
+        AUDIO_FORMAT
+            .buildUpon()
+            .setRoleFlags(C.ROLE_FLAG_DESCRIBES_VIDEO)
+            .setLabel("commentary")
+            .build();
+    Format unlabeledButRightRoleFlag =
+        AUDIO_FORMAT.buildUpon().setRoleFlags(C.ROLE_FLAG_CAPTION).setLabel(null).build();
+    TrackGroupArray trackGroups = wrapFormats(labeledButWrongRoleFlag, unlabeledButRightRoleFlag);
+    RendererCapabilities[] rendererCapabilities =
+        new RendererCapabilities[] {ALL_AUDIO_FORMAT_SUPPORTED_RENDERER_CAPABILITIES};
+
+    trackSelector.setParameters(
+        defaultParameters
+            .buildUpon()
+            .setPreferredAudioRoleFlags(C.ROLE_FLAG_CAPTION)
+            .setPreferredAudioLabels("commentary")
+            .build());
+    TrackSelectorResult result =
+        trackSelector.selectTracks(rendererCapabilities, trackGroups, periodId, TIMELINE);
+
+    assertFixedSelection(result.selections[0], trackGroups, unlabeledButRightRoleFlag);
+  }
+
   /**
    * Tests that the track selector will select a group with a single video track with a 'reasonable'
    * frame rate instead of a larger groups of tracks all with lower frame rates (the larger group of
@@ -3529,9 +3728,11 @@ public final class DefaultTrackSelectorTest {
             /* viewportHeight= */ 9,
             /* viewportOrientationMayChange= */ true)
         .setPreferredVideoMimeTypes(MimeTypes.VIDEO_AV1, MimeTypes.VIDEO_H264)
+        .setPreferredVideoLabels("video_label_1", "video_label_2")
         // Audio
         .setPreferredAudioLanguages("zh", "jp")
         .setPreferredAudioRoleFlags(C.ROLE_FLAG_COMMENTARY)
+        .setPreferredAudioLabels("audio_label_1", "audio_label_2")
         .setMaxAudioChannelCount(10)
         .setMaxAudioBitrate(11)
         .setExceedAudioConstraintsIfNecessary(false)
@@ -3547,6 +3748,7 @@ public final class DefaultTrackSelectorTest {
         .setPreferredTextRoleFlags(C.ROLE_FLAG_CAPTION)
         .setSelectUndeterminedTextLanguage(true)
         .setIgnoredTextSelectionFlags(C.SELECTION_FLAG_AUTOSELECT)
+        .setPreferredTextLabels("text_label_1", "text_label_2")
         // General
         .setForceLowestBitrate(false)
         .setForceHighestSupportedBitrate(true)